Refactor services: Move functionality from su_services to io module and update imports
