+++ docs/plans/sqlalchemy_to_flask_sqlalchemy_migration.md
@@ -906,7 +906,7 @@ def _register_error_handlers(app: Flask) -> None:
 
 
 def _register_context_processors(app: Flask) -> None:
-    from .su_services.users_service import context_data
+    from .io.users_service import context_data
     from .utils import get_status_class, short_url
 
     @app.context_processor
