+++ src/main_app/jobs_workers/admin_jobs_workers/download_main_files/download_helper.py
@@ -9,7 +9,7 @@ from typing import Any
 
 import requests
 
-from ....api_services import download_commons_file_core
+from ....api_services.files_service import download_commons_file_core
 
 logger = logging.getLogger(__name__)
