+++ src/main_app/public/shared_jobs_routes.py
@@ -26,7 +26,7 @@ from ..jobs_workers.jobs_worker import (
     start_job_form,
 )
 from ..jobs_workers.objects import JobData
-from ..su_services import load_job_result
+from ..io import load_job_result
 from .auth.utils import load_user
 from .utils.routes_utils import can_run_bg_jobs, load_auth_payload
