+++ tests/unit/io/test_jobs_files_service.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import pytest
 
-from src.main_app.su_services.jobs_files_service import (
+from src.main_app.io.jobs_files_service import (
     get_jobs_data_dir,
     load_job_result,
 )
@@ -12,7 +12,7 @@ from src.main_app.su_services.jobs_files_service import (
 
 @pytest.fixture
 def temp_jobs_dir(tmp_path):
-    with patch("src.main_app.su_services.jobs_files_service.settings") as mock_settings:
+    with patch("src.main_app.io.jobs_files_service.settings") as mock_settings:
         mock_settings.paths.public_jobs_path = str(tmp_path)
         get_jobs_data_dir.cache_clear()
         yield tmp_path
