+++ tests/unit/io/test_jobs_files_service1.py
@@ -9,7 +9,7 @@ import pytest
 
 from src.main_app.db.models.jobs import JobRecord
 from src.main_app.jobs_workers.utils import generate_result_file_name
-from src.main_app.su_services.jobs_files_service import (
+from src.main_app.io.jobs_files_service import (
     get_jobs_data_dir,
     load_job_result,
     save_job_result_by_name,
@@ -19,7 +19,7 @@ from src.main_app.su_services.jobs_files_service import (
 def test_save_job_result(tmp_path, monkeypatch):
     """Test saving a job result to a JSON file."""
     # Mock get_jobs_data_dir to use tmp_path
-    monkeypatch.setattr("src.main_app.su_services.jobs_files_service.get_jobs_data_dir", lambda: tmp_path)
+    monkeypatch.setattr("src.main_app.io.jobs_files_service.get_jobs_data_dir", lambda: tmp_path)
 
     job = JobRecord(id=1, job_type="collect_templates_data", status="pending")
 
@@ -80,7 +80,7 @@ def test_get_jobs_data_dir_not_configured(monkeypatch: pytest.MonkeyPatch):
 
     mock_settings = SimpleNamespace(paths=SimpleNamespace())
 
-    monkeypatch.setattr("src.main_app.su_services.jobs_files_service.settings", mock_settings)
+    monkeypatch.setattr("src.main_app.io.jobs_files_service.settings", mock_settings)
     get_jobs_data_dir.cache_clear()
 
     with pytest.raises(RuntimeError, match="jobs_path configuration is required for job result storage"):
@@ -97,7 +97,7 @@ def test_get_jobs_data_dir_creates_directory(tmp_path, monkeypatch: pytest.Monke
     assert not jobs_dir.exists()
 
     mock_settings = SimpleNamespace(paths=SimpleNamespace(jobs_path=str(jobs_dir)))
-    monkeypatch.setattr("src.main_app.su_services.jobs_files_service.settings", mock_settings)
+    monkeypatch.setattr("src.main_app.io.jobs_files_service.settings", mock_settings)
     get_jobs_data_dir.cache_clear()
 
     result = get_jobs_data_dir()
@@ -111,7 +111,7 @@ def test_save_job_result_with_datetime(tmp_path, monkeypatch: pytest.MonkeyPatch
     """Test saving a job result with datetime objects (default serialization)."""
     from datetime import datetime
 
-    monkeypatch.setattr("src.main_app.su_services.jobs_files_service.get_jobs_data_dir", lambda: tmp_path)
+    monkeypatch.setattr("src.main_app.io.jobs_files_service.get_jobs_data_dir", lambda: tmp_path)
 
     job = JobRecord(id=1, job_type="test_job", status="pending")
