+++ tests/unit/shared/auth/test_current_user.py
@@ -1,6 +1,6 @@
 # ruff: noqa: F401
 """
-Unit tests for src/main_app/su_services/current_user.py module.
+Unit urrent_user.py module.
 
 Classes to test: CurrentUser
